Added new plants to Populator. New indoorCalcFrag added.
New pants added: flowers/herbs to populator.

Added photos, careTips, adjusted punctuation on some careTips.

Reformatted excel populator sheet with new info and correct image file path: ./img/ from the incorrect ./images/

created new indoorCalcFrag to calculate the indoor start date to display on the SelectedPlantsView that is independent of the CalendarButton functionality (because of how the calendarbutton needs the date to be formatted we cannot format the date on the page itself - we have to call a new function to format the indoor start date separately).

CalendarButton format:
Tue Apr 19 2022 00:00:00 GMT-0400 (Eastern Daylight Time)

IndoorCalcFrag format with just .toDateString()
Tue Apr 19 2022

Trimmed off the day of the week from displaying on the IndoorCalcFrag() on the SelectedPlantsView:

IndoorCalcFrag format with  .toDateString().replace(/^\S+\s/, "")
Apr 19 2022